// src/utils/types/ActionType.types.ts
export const enum ActionType {
  AddBoard = 'addBoard',
  DeleteBoard = 'deleteBoard',
  RenameBoard = 'renameBoard',
  AddColumn = 'addColumn',
  DeleteColumn = 'deleteColumn',
  RenameColumn = 'renameColumn',
  AddTask = 'addTask',
  RenameTask = 'renameTask',
  DeleteTask = 'deleteTask',
  DeleteProfile = 'deleteProfile',
}